LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-mouse-operations-expected.txt

Tests if calendar picker key bindings work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS selectedDate() is "2000-01-02"
PASS currentMonth() is "2000-01"
Check that hovering over an entry highlights it.
PASS selectedDate() is "1999-12-31"
PASS currentMonth() is "2000-01"
PASS selectedDate() is "2000-01-04"
PASS currentMonth() is "2000-01"
Check that navigating between months maintains selection position.
PASS selectedDate() is "2000-02-08"
PASS currentMonth() is "2000-02"
Check that moving the mouse outside the popup de-highlights entries.
PASS selectedDate() is undefined.
Check that mouse click closes the popup and sets the value.
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("date").value is "2000-02-05"
PASS successfullyParsed is true

TEST COMPLETE
